pyproject.toml
[project]
name = "unsubscriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dave Williams", email = "dave@dave.io" }
]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "lxml>=6.0.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "structlog>=25.4.0",
    "tqdm>=4.67.1",
]

[project.scripts]
unsubscriber = "unsubscriber:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
